medical-chatbot/requirements.txt
langchain==0.1.16
chainlit==1.0.200
transformers==4.37.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
huggingface_hub==0.22.0
datasets==2.17.0
torch==2.7.0
torchvision==0.22.0
chromadb==0.4.22
python-dotenv==1.0.0
tqdm==4.66.1
pydantic==2.5.3